Feedback, Login and Logout completed. Just some tweaks and change, forgot password are balace.
