+++ fbproject/fbproject/templates/about.html
@@ -1,0 +1,7 @@
+{% extends "userpage.html" %}
+
+{% block content %}
+
+<h1>About Developers:</h1>
+
+{% endblock%}
