+++ fbproject/fbproject/templates/admin/add staff.html
@@ -1,0 +1,45 @@
+{% extends "adminpage.html" %}
+
+{% block content %}
+
+{% with messages = get_flashed_messages() %}
+  {% if messages %}
+     
+    <script type="text/javascript">
+      var m={{ messages|safe }};
+      for(var i=0;i<m.length;i++)
+      {
+        alert(m[i]);
+      }
+    </script>
+    
+
+
+  {% endif %}
+{% endwith %}
+
+
+<h1>Add an Helping hand</h1>
+
+<div class="content-section">
+<form class="form-horizontal"  method="POST">
+  <fieldset>
+    
+    <legend>Add Details</legend>
+     {{form.csrf_token }}
+      
+      {{form.name.label(class = "form-control-label")}} 
+      {{form.name(class="form-control form-control-lg")}}
+      {{form.email.label(class = "form-control-label")}} 
+      {{form.email(class="form-control form-control-lg")}}
+      {{form.password.label(class = "form-control-label")}} 
+      {{form.password(class = "form-control")}}
+
+      
+      
+      <br>  <input type="submit" class="btn btn-primary" value = "Add Member">
+      <a class="btn btn-primary-outline" href="{{url_for('home')}}">Cancel</a>
+  </fieldset>
+</form>
+</div>
+{% endblock %}
