+++ fbproject/fbproject/templates/admin/edit staff.html
@@ -1,0 +1,41 @@
+\]{% extends "adminpage.html" %}
+
+{% block content %}
+
+
+
+<h1></h1>
+<form class="form-horizontal"  method="POST">
+
+ <div class="form-group">
+   <label class="control-label col-sm-2" for="pwd">Name:</label>
+   <div class="col-sm-6">
+     <input type="text" class="form-control" id="name" name="name" >
+   </div>
+ </div>
+ 
+
+<div class="form-group">
+   <label class="control-label col-sm-2" for="pwd">Email</label>
+   <div class="col-sm-6">
+     <input type="text" class="form-control" id="email" name="email" >
+   </div>
+ </div>
+
+ <div class="form-group">
+   <label class="control-label col-sm-2" for="pwd">Password:</label>
+   <div class="col-sm-6">
+     <input type="text" class="form-control" id="password" name="password" >
+   </div>
+ </div>
+ 
+ 
+
+<div class="form-group">
+    <div class="col-sm-offset-2 col-sm-10">
+      <button type="submit" class="btn btn-default">Submit</button>
+    </div>
+  </div>
+</form>
+{% endblock %}
+
