+++ fbproject/fbproject/templates/admin/forgot pswd.html
@@ -1,0 +1,17 @@
+{% extends "userpage.html" %}
+
+{% block content %}
+
+
+<div class="content-section">
+<form class="form-horizontal"  method="POST">
+  <fieldset>
+    <legend>Enter email to Change the Password</legend>
+     {{form.csrf_token}}
+      {{form.email.label(class = "form-control-label")}} 
+      {{form.email(class="form-control form-control-lg", placeholder = "asdf@dedd.com")}}
+      <br>  {{form.submit(class = "btn btn-primary")}}
+  </fieldset>
+</form>
+</div>
+{% endblock %}
