+++ fbproject/fbproject/templates/admin/login.html
@@ -1,0 +1,23 @@
+{% extends "userpage.html" %}
+{% block content %}
+<h1>Login  page</h1>
+<h4>This is strictly restricted for admins only</h4>
+<form   method="POST">
+      {{form.csrf_token}}
+
+   {{form.username.label(class = "form-control-label")}}
+   {{form.username(class = "form-control")}}
+
+   {{form.password.label(class = "form-control-label")}}
+   {{form.password(class = "form-control")}}
+   <br>
+   {{form.remember}} {{form.remember.label(class = "form-contol-label")}}
+    <br>     
+   {{form.login(class = "btn btn-primary")}}
+
+</form>
+<br>  
+
+<a href="{{url_for('forgotpassword')}}">Forgot Password</a>
+</div>
+{% endblock %}
