+++ fbproject/fbproject/templates/feedback page.html
@@ -1,0 +1,26 @@
+{% extends 'userpage.html '%}
+{% block content %}
+
+<h1>{{event.name}}</h1>
+<h3>We Thank you for your visit and valuable feedback.</h3>
+<br><br>
+
+
+	<form method="post">
+		<legend>Please leave your details</legend>
+		{{form.by.label(class = "form-control-label")}}
+		{{form.by(class = "form-control")}}
+		{{form.email.label(class = "form-control-label")}}
+		{{form.email(class = "form-control")}}
+		 	<div class="input-group">
+      <div class="input-group-prepend">
+        {{form.msg.label(class = "input-group-text")}}
+      </div>
+        {{form.msg(class="form-control form-control-lg")}}
+      </div>
+      <br><br>
+      <input type="submit" value="Submit Feedback" class = "btn btn-primary">
+      <a href="/" class = "btn btn-primary-outline">Cancel</a>
+	</form>
+
+{% endblock %}
