+++ fbproject/fbproject/templates/future events.html
@@ -1,0 +1,3 @@
+{% extends 'userhome.html' %}
+
+{% block heading %} Events Upcoming: {% endblock %}
