+++ fbproject/fbproject/templates/userpage.html
@@ -1,0 +1,56 @@
+
+<!DOCTYPE html>
+<html>
+<head>
+<title>my project</title>
+<meta charset="utf-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
+  <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
+    <script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.0/js/bootstrap.min.js"></script>
+  <head>
+<body>
+<nav class="navbar navbar-inverse">
+   
+   <div class="navbar-header">
+       <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#myNavbar">
+       <span class="icon-bar"></span>
+       <span class="icon-bar"></span>
+       <span class="icon-bar"></span>
+       </button>
+       <a href="/" class="navbar-brand" style="color:white">Project</a>
+   </div>
+  <div class="collapse navbar-collapse" id="myNavbar">
+     <ul class="nav navbar-nav">
+       <li><a href="{{url_for('future_events')}}" style="color:white">Future Events</a></li>
+       
+       <li><a href="{{url_for('past_events')}}" style="color:white"> Past Events</a></li>
+       <li><a href="{{url_for('about')}}" style="color:white">About</a></li>
+       <li><a href="{{url_for('login')}}" style="color:white">Log In
+      
+       <p class="glyphicon glyphicon-lock" style="color:white"></p></a></li>
+       
+      
+      
+       
+       
+       
+       
+               
+     </ul>
+  </div>
+   
+</nav>
+<div class="container">
+{% block content %}
+
+{% endblock %}
+</div>
+<!-- <div class="container-fluid bg-info">
+<center><h1>Footer</h1></center>
+</div> -->
+<script type="text/javascript" src="https://code.jquery.com/jquery-3.5.1.js"></script>
+<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/js/bootstrap.min.js"></script>
+</body>
+</html>
+
