--- fbproject/templates/about.html
@@ -1,7 +1,0 @@
-{% extends "userpage.html" %}
-
-{% block content %}
-
-<h1>About Developers:</h1>
-
-{% endblock%}
