--- fbproject/templates/eventcreation.html
@@ -1,53 +1,0 @@
-{% extends "adminpage.html" %}
-
-{% block content %}
-
-{% with messages=get_flashed_messages() %}
-  {% if messages %}
-    <script type="text/javascript">
-      var m={{ messages|safe }};
-      for(var i=0;i<m.length;i++)
-      {
-        alert(m[i]);
-      }
-    </script>
-  {% endif %}
-{% endwith %}
-
-
-<div class="content-section">
-<form class="form-horizontal"  method="POST">
-  <fieldset>
-    <legend>Create an Event</legend>
-     {{form.csrf_token}}
-      {{form.eventname.label(class = "form-control-label")}} 
-      {{form.eventname(class="form-control form-control-lg")}}
-      {{form.course.label(class = "form-control-label")}}
-      {{form.course(class = "form-control")}}
-      <small>(Please enter course names separated by a comma)</small>
-      <br>
-      {{form.regopen.label(class = "form-control-label")}} 
-      {{form.regopen(class="form-control form-control-lg")}}
-      {{form.regclose.label(class = "form-control-label")}} 
-      {{form.regclose(class="form-control ")}} 
-      {{form.place.label(class = "form-control-label")}} 
-      {{form.place(class="form-control form-control-lg")}}
-      {{form.date.label(class = "form-control-label")}} 
-      {{form.date(class="form-control form-control")}}
-      {{form.status.label(class = "form-control-label")}} 
-      {{form.status(class="form-control form-control")}}
-      
-      <div class="input-group">
-      <div class="input-group-prepend">
-        {{form.description.label(class = "input-group-text")}}
-      </div>
-        {{form.description(class="form-control form-control-lg")}}
-      </div>
-      {{form.poster.label(class = "form-control-label")}} {{form.poster}}
-      <small>Add jpeg or jpg files only.</small><br>
-      <input type="submit" value = "Add Event" class="btn btn-primary">
-      <a class="btn btn-primary-outlin">Cancel</a>
-    </fieldset>
-</form>
-</div>
-{% endblock %}
