--- fbproject/templates/eventedit.html
@@ -1,40 +1,0 @@
-{% extends "adminpage.html" %}
-
-{% block content %}
-
-
-
-<h1>Edit event</h1>
-<form class="form-horizontal"  method="POST">
-
- <div class="form-group">
-   <label class="control-label col-sm-2" for="pwd">Event name:</label>
-   <div class="col-sm-6">
-     <input type="text" class="form-control" id="eventname" name="eventname" placeholder="Enter ur choice">
-   </div>
- </div>
- 
-
-    <div class="form-group">
-    <label class="control-label col-sm-2" for="pwd">Courses:</label>
-    <div class="col-sm-6">
-    <div class="radio_inline">
-        <label class="checkbox-inline"><input type="checkbox" name="courses" value="Python">Python</label>
-<label class="checkbox-inline"><input type="checkbox" name="courses" value="ML">ML</label>
-<label class="checkbox-inline"><input type="checkbox" name="courses" value="DL">DL</label>
-<label class="checkbox-inline"><input type="checkbox" name="courses" value="RPA">RPA</label>
-<label class="checkbox-inline"><input type="checkbox" name="courses" value="Networking">Networking</label>
-<label class="checkbox-inline"><input type="checkbox" name="courses" value="Meanstack">Meanstack</label>
-<label class="checkbox-inline"><input type="checkbox" name="courses" value="WT">WT</label>
-    </div>
-    </div>
-  </div>
-
-<div class="form-group">
-    <div class="col-sm-offset-2 col-sm-10">
-      <button type="submit" class="btn btn-default">Submit</button>
-    </div>
-  </div>
-</form>
-{% endblock %}
-
