--- fbproject/templates/login.html
@@ -1,22 +1,0 @@
-{% extends "userpage.html" %}
-{% block content %}
-<h1>Login  page</h1>
-<h4>This is strictly restricted for admins only</h4>
-<form   method="POST">
-      {{form.csrf_token}}
-
-   {{form.username.label(class = "form-control-label")}}
-   {{form.username(class = "form-control")}}
-
-   {{form.password.label(class = "form-control-label")}}
-   {{form.password(class = "form-control")}}
-
-    <br>     
-   {{form.login(class = "btn btn-primary")}}
-
-</form>
-<br>  
-
-<a href="{{url_for('forgotpassword')}}">Forgot Password</a>
-</div>
-{% endblock %}
