--- fbproject/templates/staffhome.html
@@ -1,22 +1,0 @@
-{% extends "userpage.html" %}
-
-{% block content %}
-
-
-<div class="content-section">
-<form class="form-horizontal"  method="POST">
-  <fieldset>
-    <legend>Register for a Great Time</legend>
-     {{form.hidden_tag()}}
-      {{form.name.label(class = "form-control-label")}} 
-      {{form.name(class="form-control form-control-lg")}}
-      {{form.email.label(class = "form-control-label")}} 
-      {{form.email(class="form-control form-control-lg")}}
-      {{form.password.label(class = "form-control-label")}} 
-      {{form.password(class="form-control form-control-lg")}}
-      
-      <br>  {{form.submit(class = "btn btn-primary")}}
-  </fieldset>
-</form>
-</div>
-{% endblock %}
