--- fbproject/templates/view.html
@@ -1,85 +1,0 @@
-{% extends "adminpage.html" %}
-
-{% block content %}
-
-{% with messages = get_flashed_messages() %}
-  {% if messages %}
-     
-    <script type="text/javascript">
-      var m = {{ "messages|safe" }}
-      for(var i=0;i<m.length;i++)
-      {
-        alert(m[i]);
-
-      }
-    </script>
-    
-
-
-  {% endif %}
-{% endwith %}
-
-<style type="text/css">
-  #lastrow {
-    display: flex;
-    align-items: center;
-    justify-content: space-around;
-  }
-  th#aaa {
-    text-align: center;
-  }
-</style>
-
-
-
-
-
-
-<table id="example" class="display" style="width:100%">
-        <thead>
-            <tr>
-                <th>S.No</th>
-
-                <th>event</th>
-                <th>courses</th>
-                <th id="aaa">Action</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for d in data %}
-            <tr>
-                <td>{{loop.index}}</td>
-                <td>{{d[1]}}</td>
-                <td>{{d[2]}}</td>
-                
-                  <td id= "lastrow">
-                  <form action="{{url_for('edit')}}" method="POST">
-                  <button value={{d[0]}} name="edit">Edit</button>
-                  </form>
-                  <form action="{{url_for('delete')}}" method="POST">
-                  <button value={{d[1]}} name="delete">Delete</button>
-                  </form>
-                </td>
-
-
-            </tr>
-            {% endfor %}
-
-        </tbody>
-</table>
-
-
-<script type="text/javascript">
-
-  $(document).ready(function() {
-    $('#example').DataTable( {
-        dom: 'Bfrtip',
-        buttons: [
-            'copy', 'csv', 'excel', 'pdf', 'print'
-        ]
-    } );
-} );
-
-</script>
-
-{% endblock %}
